[clang-tidy] Add new alias 'bugprone-unchecked-string-to-number-conversion' for 'cert-err34-c'
